Fix some JS and React warnings.
Gatsby's Link should be used for interanl linking. Some links passed to this component were exterenal. The custom Link component is built to handle both scenerios.

In some cases, the title passed to the Card component wasn't set. This title was used as the Card Avatar alt text value. This fixes the warning of not passing an alt text value.