Link featured star icon on for speakers to the featured speakers archive.
Initially, there was only an archive for featured talks. The featured icon was linked to the talks archive. Since then, the icon is also used to display featured Speakers too. The issue was the featured icon was always linking to the Featured Talks archive. This fix, though maybe not the most performant, parses the URL and pulls the archive context from there.